extract format field into own component
